Refactor engine experiment module organization.
This provides a bit better origanization of types and, in particular,
the `example_planners.py` provides a home for all experiment
addressables, planners and tasks along with a graph and scheduler setup
function for test files to use.
